+++ api/aks/main.go
@@ -7,10 +7,11 @@ package aks
 
 import (
 	"fmt"
-	"github.com/kubesimplify/ksctl/src/api/payload"
 	"os"
 	"runtime"
 	"strings"
+
+	"github.com/kubesimplify/ksctl/api/payload"
 )
 
 func getKubeconfig(params ...string) string {