+++ cli/cmd/deleteClusterLocal.go
@@ -8,7 +8,8 @@ Kubesimplify
 */
 import (
 	"fmt"
-	"github.com/kubesimplify/ksctl/src/api/local"
+
+	"github.com/kubesimplify/ksctl/api/local"
 	"github.com/spf13/cobra"
 )