+++ cli/main.go
@@ -3,7 +3,7 @@ Copyright © 2022 NAME HERE <EMAIL ADDRESS>
 */
 package main
 
-import "github.com/kubesimplify/ksctl/src/cli/cmd"
+import "github.com/kubesimplify/ksctl/cli/cmd"
 
 func main() {
 	cmd.Execute()