+++ install.sh
@@ -4,21 +4,15 @@
 
 #wget -O ~URL_OF_BINARY~
 
-cd src/cli || echo -e "\033[31;40mPath couldn't be found\033[0m\n"
+cd ./cli || echo -e "\033[31;40mPath couldn't be found\033[0m\n"
 # Check if sudo access
 go get -d
 go build -v -o ksctl .
-cd - || echo -e "\033[31;40mFailed to move to previous directory\033[0m\n"
-chmod +x src/cli/ksctl
+chmod +x ksctl
 
-
-sudo mv -v src/cli/ksctl /usr/local/bin/ksctl
-#cp -v ./ksctl /usr/local/bin
+sudo mv -v ksctl /usr/local/bin/ksctl
 
 # Setup the configurations dir
-#sudo mkdir /etc/ksctl
-
-# Setup the ~/.ksctl
 mkdir -p ${HOME}/.ksctl/cred
 touch ${HOME}/.ksctl/cred/aws
 touch ${HOME}/.ksctl/cred/azure
@@ -31,3 +25,5 @@ mkdir ${HOME}/.ksctl/config/aws
 mkdir ${HOME}/.ksctl/config/local
 
 echo -e "\033[32;40mINSTALL COMPLETE\033[0m\n"
+
+cd - || echo -e "\033[31;40mFailed to move to previous directory\033[0m\n"